Abstract
Convolution Neural Network (CNN) have been widely used in automatic image classification systems. Image classification requires generation of features capable of detecting image patterns informative of group identity. A variety of image data sets are available to test the performance of different types of CNN’s. The commonly found benchmark datasets for evaluating the performance of a convolution neural network are CIFAR -10, CIFAR-100 and MNIST image data sets. One of the most popular data sets, CIFAR-10 has been taken to test the performance of the network. The main objective of this study is to get a better test accuracy with reduced overfitting which was achieved with a deeper network. The deep learning models are implemented with the use of Keras library available for Python programming language. An accuracy of around 96.99% is expected after running 50 epochs. 
 
•	Images can be classified and kept in separate folders which will avoid confusion and will be easier to locate. 
•	Automatic face recognition and object recognition can be used for classifying the images automatically. 
•	A lot of tasks can be automated using image recognition like processing cheques in banks, computer vision, space exploration, processing signal images etc. 
•	Various application in industrial, scientific and medical.